WebServer: split to its own file
Moved WebServer and its dependencies as-is to a new file.
